+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.4]
+- Always have a default HTTP client even if no client is set.
+- Always have a default Message Factory even if no client is set.
+- Set API wrapper for Validation and Telemetry classes even if no client is set.
+
 ## [2.0.3]
 - Fix use of some methods with multiple clients.